+++ src/Raven.Server/Documents/Handlers/IndexHandler.cs
@@ -996,7 +996,7 @@ namespace Raven.Server.Documents.Handlers
                         {
                             Value = 0
                         };
-                        var docs = Database.DocumentsStorage.GetDocuments(context, listOfIds, 0, int.MaxValue, _);
+                        var docs = Database.DocumentsStorage.GetDocuments(context, listOfIds, 0, long.MaxValue, _);
                         foreach (var doc in docs)
                         {
                             if (doc.TryGetMetadata(out var metadata) && metadata.TryGet(Constants.Documents.Metadata.Collection, out string collectionStr))
